+++ src/cinegraph/common/error_messages/__init__.py
@@ -8,6 +8,7 @@ from cinegraph.common.error_messages.corpus_ingestion import (
 )
 from cinegraph.common.error_messages.evaluation import EvaluationErrorMessages
 from cinegraph.common.error_messages.grounded_answer import GroundedAnswerErrorMessages
+from cinegraph.common.error_messages.ingestion_jobs import IngestionJobErrorMessages
 from cinegraph.common.error_messages.media_action import MediaActionErrorMessages
 from cinegraph.common.error_messages.media_provider import MediaProviderErrorMessages
 from cinegraph.common.error_messages.netflix_history import NetflixHistoryErrorMessages
@@ -44,4 +45,5 @@ __all__ = [
     "TranscriptErrorMessages",
     "WatchErrorMessages",
     "WorkflowErrorMessages",
+    "IngestionJobErrorMessages",
 ]